Issue 3985 - Separated advanced from canvas in IB settings
