+++ src/blocks/svg-icon-maxi/data.js
@@ -19,7 +19,10 @@ import {
 	getBorderStyles,
 	getSVGStyles,
 } from '../../extensions/styles/helpers';
-import { getCanvasSettings } from '../../extensions/relations';
+import {
+	getCanvasSettings,
+	getAdvancedSettings,
+} from '../../extensions/relations';
 import transitionDefault from '../../extensions/styles/transitions/transitionDefault';
 
 /**
@@ -265,7 +268,8 @@ const interactionBuilderSettings = {
 			target: ' .maxi-svg-icon-block__icon',
 		},
 	],
-	canvas: getCanvasSettings({ name, customCss }),
+	canvas: getCanvasSettings({ name }),
+	advanced: getAdvancedSettings({ customCss }),
 };
 
 const data = {
