+++ src/blocks/text-maxi/data.js
@@ -12,7 +12,10 @@ import {
 	getAlignmentTextStyles,
 	getTypographyStyles,
 } from '../../extensions/styles/helpers';
-import { getCanvasSettings } from '../../extensions/relations';
+import {
+	getCanvasSettings,
+	getAdvancedSettings,
+} from '../../extensions/relations';
 import transitionDefault from '../../extensions/styles/transitions/transitionDefault';
 
 /**
@@ -201,7 +204,8 @@ const interactionBuilderSettings = {
 			target: contentClass,
 		},
 	],
-	canvas: getCanvasSettings({ name, customCss }),
+	canvas: getCanvasSettings({ name }),
+	advanced: getAdvancedSettings({ customCss }),
 };
 
 const data = {
