+++ src/blocks/video-maxi/data.js
@@ -1,7 +1,10 @@
 /**
  * Internal dependencies
  */
-import { getCanvasSettings } from '../../extensions/relations';
+import {
+	getCanvasSettings,
+	getAdvancedSettings,
+} from '../../extensions/relations';
 import { createIconTransitions } from '../../extensions/styles';
 import { createSelectors } from '../../extensions/styles/custom-css';
 import transitionDefault from '../../extensions/styles/transitions/transitionDefault';
@@ -220,7 +223,8 @@ const transition = {
 	},
 };
 const interactionBuilderSettings = {
-	canvas: getCanvasSettings({ name, customCss }),
+	canvas: getCanvasSettings({ name }),
+	advanced: getAdvancedSettings({ customCss }),
 };
 
 const data = {
