+++ src/extensions/relations/getCanvasSettings.js
@@ -13,65 +13,23 @@ import {
 	FullSizeControl,
 	InfoBox,
 	MarginControl,
-	OpacityControl,
 	PaddingControl,
-	PositionControl,
-	TransformControl,
 } from '../../components';
 import {
 	getBlockBackgroundStyles,
 	getBorderStyles,
 	getBoxShadowStyles,
 	getMarginPaddingStyles,
-	getOpacityStyles,
-	getPositionStyles,
 	getSizeStyles,
-	getTransformStyles,
 } from '../styles/helpers';
-import {
-	getTransformCategories,
-	getTransformSelectors,
-} from '../../components/transform-control/utils';
 import { getGroupAttributes, getLastBreakpointAttribute } from '../styles';
 
 /**
  * External dependencies
  */
-import { isEmpty, isEqual, isPlainObject, pickBy } from 'lodash';
+import { isEmpty, isEqual, pickBy } from 'lodash';
 
-const getTransformControl = ({ categories, selectors }) => ({
-	label: __('Transform', 'maxi-blocks'),
-	transitionTarget: [],
-	hoverProp: (attributes, relationAttributes) =>
-		Object.entries(getGroupAttributes(attributes, 'transform')).some(
-			([attributeKey, attribute]) =>
-				relationAttributes?.[attributeKey] &&
-				isPlainObject(attribute) &&
-				Object.entries(attribute).some(
-					([objKey, obj]) =>
-						relationAttributes[attributeKey][objKey] &&
-						obj?.['hover-status']
-				)
-		),
-	attrGroupName: 'transform',
-	component: props => (
-		<TransformControl
-			{...props}
-			uniqueID={props.attributes.uniqueID}
-			depth={2}
-			selectors={getTransformSelectors(selectors, props.attributes)}
-			categories={getTransformCategories(categories, props.attributes)}
-			disableHover
-		/>
-	),
-	helper: props =>
-		getTransformStyles(
-			props.obj,
-			getTransformSelectors(selectors, props.blockAttributes)
-		),
-});
-
-const getCanvasSettings = ({ name, customCss }) => [
+const getCanvasSettings = ({ name }) => [
 	{
 		label: __('Background / Layer', 'maxi-blocks'),
 		transitionTarget: ' > .maxi-background-displayer > div',
@@ -148,22 +106,6 @@ const getCanvasSettings = ({ name, customCss }) => [
 		helper: props => getBoxShadowStyles(props),
 	},
 	{
-		label: __('Opacity', 'maxi-blocks'),
-		hoverProp: 'opacity-status-hover',
-		attrGroupName: 'opacity',
-		component: props => (
-			<OpacityControl
-				{...props}
-				opacity={getLastBreakpointAttribute({
-					target: 'opacity',
-					breakpoint: props.breakpoint,
-					attributes: getGroupAttributes(props, 'opacity'),
-				})}
-			/>
-		),
-		helper: props => getOpacityStyles(props.obj),
-	},
-	{
 		label: __('Height / Width', 'maxi-blocks'),
 		attrGroupName: 'size',
 		component: props => {
@@ -197,13 +139,6 @@ const getCanvasSettings = ({ name, customCss }) => [
 		),
 		helper: props => getMarginPaddingStyles(props),
 	},
-	{
-		label: __('Position', 'maxi-blocks'),
-		attrGroupName: 'position',
-		component: props => <PositionControl {...props} />,
-		helper: props => getPositionStyles(props.obj),
-	},
-	...getTransformControl(customCss),
 ];
 
 export default getCanvasSettings;
