+++ src/extensions/relations/index.js
@@ -1,2 +1,3 @@
 export { default as getCanvasSettings } from './getCanvasSettings';
+export { default as getAdvancedSettings } from './getAdvancedSettings';
 export { default as getHoverStatus } from './getHoverStatus';
